full migration to artifactory.v54 and outputter in cmds (#23)

+++ cmd/artif-delete-user/main.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	artifactory "github.com/lusis/go-artifactory/artifactory.v51"
+	artifactory "github.com/lusis/go-artifactory/artifactory.v54"
 	kingpin "gopkg.in/alecthomas/kingpin.v2"
 )
 
@@ -14,7 +14,11 @@ var (
 
 func main() {
 	kingpin.Parse()
-	client := artifactory.NewClientFromEnv()
+	client, clientErr := artifactory.NewClientFromEnv()
+	if clientErr != nil {
+		fmt.Printf("%s\n", clientErr.Error())
+		os.Exit(1)
+	}
 	err := client.DeleteUser(*username)
 	if err != nil {
 		fmt.Printf("%s\n", err)
